backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
python-multipart==0.0.20
PyJWT==2.10.1
cryptography==44.0.1
svix==1.66.0
